refactor: extract sandbox backend tests
