+++ scripts/rust-source-size-baseline.txt
@@ -1,7 +1,7 @@
 # Rust source files above the 1,000-line target. Keep sorted by path.
 # A file may only stay equal to this maximum or shrink with this entry tightened.
 crates/agent-engine/src/tools/sandbox.rs 3514
-crates/agent-engine/src/tools/sandbox_backend.rs 1562
+crates/agent-engine/src/tools/sandbox_backend.rs 1007
 crates/agent-engine/src/tools/sandbox_tests.rs 2976
 crates/agentfs/src/lib.rs 1151
 crates/agentfs/src/root.rs 1191
